\name{cnvGSA_input_example.RData}
\alias{cnvGSA_input_example.RData}
\alias{input}
\docType{data}
\title{cnvGSA example input}
\description{
Pre-built input used in the full workflow example in the \code{cnvGSA} vignette.
}
\details{
	The dataset contains \code{input}, a single object of class \code{CnvGSAInput} as defined in the \code{cnvGSA} package. When processed by \code{cnvGSAFisher()} -- the main function in that package -- it produces as its output an object of class \code{CnvGSAOutput} (such as the one stored in cnvGSA_output_example.RData).
}
\usage{data("cnvGSA_input_example")}
