\name{cnvGSAdata-package}
\alias{cnvGSAdata-package}
\alias{cnvGSAdata}
\docType{package}
\title{
Data used in the examples and vignette of the cnvGSA package
}
\description{
	This package contains data used in the examples and vignette of the cnvGSA package.
}
\details{
	Raw input files:
	\code{\link{cnv_AGP_demo.txt}} - CNV data
	\code{\link{enrGeneric_AGP_demo.txt}} - enrichment generic data
	\code{\link{enrGMT_AGP_demo.gmt}} - enrichment GMT data
	\code{\link{kl_gene_AGP_demo.txt}} - genes of interest data
	\code{\link{kl_loci_AGP_demo.txt}} - known loci data
	\code{\link{ph_AGP_demo.txt}} - phenotype/covariate data
	\code{\link{gene_ID_demo.txt}} - gene ID file 

	Pre-built input:
	\code{\link{cnvGSA\_input\_example.RData}}

	Saved output:
	\code{\link{cnvGSA\_output\_example.RData}}

	Pre-built Gene-set data:
	\code{\link{gs_data_example.RData}}
}
\author{Joseph Lugo \email{joseph.r.lugo@gmail.com}}
