\name{gs_data_example.RData}
\alias{gs_data_example.RData}
\alias{gsData}
\docType{data}
\title{gsData example output}
\description{
Saved gsData described in the \code{cnvGSA} vignette.
}
\details{
	The dataset contains \code{gsData}. It contains the gene-sets analyzed in the \code{cnvGSA} vignette as well as the gene set names.

	\describe{
		\item{\code{gs_all.ls}}{}
		\item{\code{gsid2name.chv}}{}
	}
	
	gs_all.ls contains the gene-sets used in the analysis and gsid2name.chv contains the gene-set names.

	See the \code{cnvGSA} vignette for complete details and discussion.
}
\usage{data("gs_data_example")}