\name{cnvGSA\_input\_example.RData}
\alias{cnvGSA\_input\_example.RData}
\alias{cnvGSA.in}
\docType{data}
\title{cnvGSA example input}
\description{
Pre-built input used in the full workflow example in the \code{cnvGSA} vignette.
}
\details{
	The dataset contains \code{cnvGSA.in}, a single object of class \code{CnvGSAInput} as defined in the \code{cnvGSA} package. When processed by \code{cnvGSAlogRegTest()} -- the main function in that package -- it produces as its output an object of class \code{CnvGSAOutput} (such as the one stored in cnvGSA_output_example.RData).
}
\usage{data("cnvGSA_input_example")}
