\name{cnvGSAdata-package}
\alias{cnvGSAdata-package}
\alias{cnvGSAdata}
\docType{package}
\title{
Data used in the examples and vignette of the cnvGSA package
}
\description{
	This package contains data used in the examples and vignette of the cnvGSA package.
}
\details{
	Raw input files:
	\code{\link{cnv.gvf}} - CNV data
	\code{\link{gsData.gmt}} - Gene-set data
	\code{\link{merge_00k_flank_hg18_refGene_jun_2011_exon.gff}} - Gene map for exons
	\code{\link{merge_00k_flank_hg18_refGene_jun_2011_transcript.gff}} - Gene map for transcripts
	\code{\link{s2class.txt}} - Sample-to-class information (for the CNV data)

	Pre-built input:
	\code{\link{cnvGSA_input_example.RData}}

	Saved output:
	\code{\link{cnvGSA_output_example.RData}}
}
\author{Robert Ziman \email{rziman@gmail.com}}
